// ui/src/generated/api/model/dtoAccountUpdateRequestRole.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */

export type DtoAccountUpdateRequestRole =
  (typeof DtoAccountUpdateRequestRole)[keyof typeof DtoAccountUpdateRequestRole];

export const DtoAccountUpdateRequestRole = {
  user: "user",
  admin: "admin",
} as const;