// ui/src/generated/api/model/getApiAdminUsersRole.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */

export type GetApiAdminUsersRole = (typeof GetApiAdminUsersRole)[keyof typeof GetApiAdminUsersRole];

export const GetApiAdminUsersRole = {
  user: "user",
  admin: "admin",
} as const;